+++ src/main/java/org/owasp/webgoat/lessons/passwordreset/TriedQuestions.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.password_reset;
+package org.owasp.webgoat.lessons.passwordreset;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.annotation.SessionScope;
