+++ src/main/resources/lessons/passwordreset/documentation/PasswordReset_mitigation.adoc
