+++ src/main/resources/lessons/passwordreset/documentation/PasswordReset_simple.adoc
