+++ src/main/resources/lessons/passwordreset/images/reset2.png
