+++ src/main/resources/lessons/passwordreset/js/password-reset-simple.js
